requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-filter==25.2
django-redis==5.4.0

psycopg2==2.9.11

drf-spectacular==0.27.2

# OpenAI / ChatGPT API
openai==2.7.2
